Get application detail screen working
