+++ packages/user/FractalCore/ui/src/lib/graphql/fractals/getGuildApplications.ts
@@ -1,29 +1,10 @@
 import { z } from "zod";
 
 import { FRACTALS_SERVICE } from "@/lib/constants";
+import { zGuildApplicationListInstance } from "@/lib/zod/attestations";
 
 import { graphql } from "@shared/lib/graphql";
-import { Account, zAccount } from "@shared/lib/schemas/account";
-import { zDateTime } from "@shared/lib/schemas/date-time";
-
-export const zGuildAttestationListInstance = z.object({
-    attester: zAccount,
-    comment: z.string(),
-    endorses: z.boolean(),
-});
-
-export const zGuildApplicationListInstance = z.object({
-    applicant: zAccount,
-    extraInfo: z.string(),
-    createdAt: zDateTime,
-    attestations: z.object({
-        nodes: zGuildAttestationListInstance.array(),
-    }),
-});
-
-export type GuildApplicationListInstance = z.infer<
-    typeof zGuildApplicationListInstance
->;
+import { Account } from "@shared/lib/schemas/account";
 
 export const getGuildApplications = async (guildAccount: Account) => {
     const res = await graphql(
